demain, lumiere coloree

+++ includes/lightning.h
@@ -6,7 +6,7 @@
 /*   By: lmartin <lmartin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/02 09:30:36 by lmartin           #+#    #+#             */
-/*   Updated: 2019/11/19 02:11:16 by lmartin          ###   ########.fr       */
+/*   Updated: 2019/11/19 03:40:04 by lmartin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,6 +22,6 @@
 float		calcul_shiny(s_lightning_vectors *l_vectors, s_light *light, float length_v, s_vector *vec_l);
 float		compute_special_lights(s_lightning_vectors *l_vectors, s_light *light, s_scene *scene, s_lstobjects *objects);
 s_vector	*type_light(s_lightning_vectors *l_vectors, s_light *light, s_scene *scene);
-float		compute_lightning(s_lightning_vectors *l_vectors, s_lstobjects *lights, s_scene *scene, s_lstobjects *objects);
+int		compute_lightning(s_lightning_vectors *l_vectors, s_lstobjects *lights, s_scene *scene, s_lstobjects *objects);
 
 #endif
